Run MAYA for unsupervised cell annotation with cell type

Run MAYA for unsupervised cell annotation with cell type

MAYA_predict_cell_types(
  expr_mat,
  modules_list = NULL,
  min_cells_pct = 0.05,
  organs = NULL,
  is_logcpm = F,
  nCores = 1,
  thr = 0,
  max_contrib = 0.5,
  compute_umap = T,
  plot_heatmap = T,
  scale_before_pca = T
)

Arguments

expr_mat

Raw count matrix. If normalized, set is_logcpm to T

modules_list

List with cell types associated with their markers. If NULL, PanglaoDB will be loaded.

min_cells_pct

Numeric between 0 and 1, minimum percentage of cells that should be above informativity threshold to consider activity score interesting.

organs

Character to specify if list for a specific organ should be loaded. One of ("Pancreas","Brain","Lungs","Heart","Liver","Adrenal glands","GI tract","Reproductive","Kidney","Zygote","Thyroid","Embryo","Bone","Skin","Mammary gland","Eye","Olfactory system","Oral cavity","Thymus","Placenta","Urinary bladder"). "All" is another possible value, the 147 PanglaoDB cell types will be loaded.

is_logcpm

Set to TRUE if data already normalized

nCores

Number of cores to use. Set to 1 if working in a Windows environment, otherwise you can use the function detectCores() to find out how many cores are available.

thr

Minimum average activity by cluster required to be assigned an annotation. Default 0, increase to be more stringent on assignation confidence.

max_contrib

Numeric between 0 and 1, representing the maximum contribution to a mode allowed for a gene. Can influence how stringent mode selection is.

plot_heatmap

Set to False to disable automatic display of activity matrix as a heatmap.

Value

List containing cell type annotation, matrix of average score by Leiden cluster and cell type used for attribution, cell annotation with Leiden clusters, activity matrix and UMAP computed on activity matrix.